added dgn and nac

+++ torchrl/utils/schedules.py
@@ -12,9 +12,9 @@ def linear_schedule(initial_value, final_value, final_step, initial_step=0):
     def get(step):
         step = to_np(step)
         if step <= final_step:
-            return decay_rate * (step - initial_step) + initial_value
+            return float(decay_rate * (step - initial_step) + initial_value)
         else:
-            return final_value
+            return float(final_value)
 
     return get
